--- trackpack/routes/routes.py
@@ -1,4 +1,0 @@
-#functions for routes specified in ../run.py
-
-def hello_world():
-    return "Hello World!"
